Move LookForStringInLog() from e2e framework
Both LookForStringInLog() and LookForStringInPodExec() call RunKubectlOrDie()
internally, they seem like wrapper functions for RunKubectlOrDie().
We have long-term goal for moving kubectl related functions into e2ekubectl.
So this moves the above functions for the goal as a small step.
